refactoring code to make it work with the rest of the stuff
