+++ app/src/typings/SC.ts
