+++ app/src/typings/common.ts
@@ -1,0 +1,5 @@
+import { Track } from './SC';
+
+export interface EnhancedTrack extends Track {
+  index: number;
+}
